Make `ST::Page::TablePopulator` actually support "recreate" as an *OPTION*.
Previously, it would *always* purge the Page table before re-populating it with
data.  Made the `--recreate` option in `st-populate-page-table` moot; even if
you didn't `--recreate` on the command line, `ST::Page::TablePopulator` would
still purge the Page table first.  Purging of the Page table would then cascade
through to other DB tables and cause a domino effect of things being wiped from
existence (e.g. Events).

This commit adds support for "recreate" as an option in
`ST::Page::TablePopulator`, which now allows for `st-populate-page-table` to be
run to just "fill in the missing blanks"; it'll add pages that were present on
disk but missing in the DB, but leave existing data in the DB alone.

Before committing, a search was done through our entire code base to find out
what made use of either `ST::Page::TablePopulator` or `st-populate-page-table`,
so that they could be updated to preserve their existing "recreate/purge"
behaviour.  Only *one* piece of code was found across our various SVN repos that
needed updating; `Socialtext::Workspace::Importer`.

During the above-mentioned search, it was found that the
`26-page-metadata-to-db` migration uses `ST::Page::TablePopulator`, *BUT* that
migration also has its own custom copy of `lib/` that it uses (so its *not*
using this updated version, ever).

To accompany this update, a pair of unit tests have been added to verify that
the "recreate" option works as intended (by purging the DB first if recreate=>1,
and by leaving the DB alone if recreate=>0).


git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@11941 593ce9bc-2028-0410-bb98-c566b643a948

+++ nlw/lib/Socialtext/Page/TablePopulator.pm
@@ -32,6 +32,8 @@ sub new {
 
 sub populate {
     my $self = shift;
+    my %opts = @_;
+    my $recreate = $opts{recreate};
     my $workspace = $self->{workspace};
     my $workspace_name = $self->{workspace_name};
 
@@ -39,11 +41,16 @@ sub populate {
     my $already_in_txn = sql_in_transaction();
     sql_begin_work() unless $already_in_txn;
     eval {
-        sql_execute(
-            'DELETE FROM page WHERE workspace_id = ?',
-            $workspace->workspace_id,
-        );
+        # if we're recreating the workspace from scratch, clean it out *first*
+        if ($recreate) {
+            sql_execute(
+                'DELETE FROM page WHERE workspace_id = ?',
+                $workspace->workspace_id,
+            );
+        }
 
+        # Grab all the Pages in the Workspace, figure out which ones we need
+        # to add to the DB, then add them all.
         my $workspace_dir
             = Socialtext::Paths::page_data_directory($workspace_name);
         my $hub = $self->{hub}
@@ -80,6 +87,13 @@ sub populate {
                 warn "Error populating $workspace_name: $@";
                 next PAGE;
             }
+
+            # Skip this page if it already exists in the DB
+            my $page_exists_already = page_exists_in_db(
+                workspace_id => $workspace_id,
+                page_id      => $page->{page_id},
+            );
+            next PAGE if $page_exists_already;
 
             # Add this page (and its tags) to the list of things to add to the
             # DB.
@@ -216,6 +230,21 @@ EOT
             . join(', ', @$row) . ') - '
             . $sth->errstr;
     }
+}
+
+sub page_exists_in_db {
+    my %opts = @_;
+    my $ws_id   = $opts{workspace_id} || die "workspace_id is required";
+    my $page_id = $opts{page_id}      || die "page_id is required";
+
+    my $exists_already = sql_singlevalue( q{
+        SELECT true
+          FROM page
+         WHERE workspace_id = ?
+           AND page_id = ?
+        }, $ws_id, $page_id
+    );
+    return defined $exists_already ? 1 : 0;
 }
 
 sub load_original_revision {
